4.21 complete!

+++ controllers/blogs.js
@@ -56,8 +56,19 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const token = getTokenFrom(request);
+  const blog = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  } else if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response
+      .status(401)
+      .json({ error: "blogs can only be deleted by user that posted them!" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
